Different algorithm to parse inline text
The algorithm is based on progressive replacement of parts of the text
by numbered placeholders that represent the HTML fragments that
replace those parts. When the parsing is done, the placeholders are
replaced by the corresponding HTML fragments.

All tabs are expanded, rather than only the tabs at the star of a
line.

Correctly omit empty lines at the end of indented code blocks.

Do not end a fenced code block with a newline if the block is
otherwise empty.

Escape backslashes in URLs (as "%5C"), because browsers otherwise turn
them into forward slashes. Also escape "[" and "]", because the
examples in GitHub's spec do so and this makes it easier to compare
the results to those examples.